LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/seeking/seek-to-max-value-expected.txt


FAIL seek to Number.MAX_VALUE assert_equals: currentTime after setting expected 5.057551020408163 but got 1.7976931348623157e+308
